feat: remove dead code
