+++ src/component/strategies/StrategiesList/AddStrategyButton/AddStrategyButton.tsx
