+++ src/component/strategies/StrategiesList/PredefinedBadge/PredefinedBadge.tsx
