--- src/component/strategies/StrategiesList/StrategiesList.styles.ts
@@ -1,14 +1,0 @@
-import { makeStyles } from 'tss-react/mui';
-
-export const useStyles = makeStyles()(theme => ({
-    listItem: {
-        padding: '0',
-        ['& a']: {
-            textDecoration: 'none',
-            color: 'inherit',
-        },
-        '&:hover': {
-            backgroundColor: theme.palette.grey[200],
-        },
-    },
-}));
